Change get-default implementation
In situations where the aware of components are being used across multiple
classes, it can be a bit slow having to ask for the facade-roots, which in
term results in IoC calls. Furthermore, the default "get" actually results
in 2 calls to the "getDefault" methods, which in it self is slow.

Therefore, since most of these components are based on singleton instances
in Laravel, it makes sense to jsut cache their defaults, which should
increase the performance a lot - if they are used across multiple classes.